+++ app/src/test/java/org/simple/clinic/home/overdue/search/OverdueSearchInitTest.kt
@@ -12,11 +12,23 @@ class OverdueSearchInitTest {
   fun `when overdue search screen is created, then load search history and selected overdue appointment ids`() {
     val defaultModel = OverdueSearchModel.create()
 
-    InitSpec(OverdueSearchInit())
+    InitSpec(OverdueSearchInit(false))
         .whenInit(defaultModel)
         .then(assertThatFirst(
             hasModel(defaultModel),
             hasEffects(LoadOverdueSearchHistory, LoadSelectedOverdueAppointmentIds)
+        ))
+  }
+
+  @Test
+  fun `when overdue search screen is created and overdue search v2 is enabled, then load villages and patient names`() {
+    val defaultModel = OverdueSearchModel.create()
+
+    InitSpec(OverdueSearchInit(true))
+        .whenInit(defaultModel)
+        .then(assertThatFirst(
+            hasModel(defaultModel),
+            hasEffects(LoadOverdueSearchHistory, LoadSelectedOverdueAppointmentIds, LoadVillageAndPatientNames)
         ))
   }
 }