+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Implement query to fetch village and patient names in the facility
 - Add custom text field component with chip input
 - Display search suggestions after loading in overdue search screen
+- Load village and patient names when overdue search screen is created
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-25-8342